backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF 生成和处理
reportlab==4.0.7

# 文档处理
python-docx==1.1.0
markdown==3.5.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件处理
aiofiles==23.2.1
pillow==10.1.0
aiohttp==3.9.1

# 数学公式处理
matplotlib==3.7.2
numpy==1.24.3
